+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 
 ### Fixed
 
+- Confidence measure that computes the distance LR / RL in cross checking method.
 - Creation of mask dataset in `read_img` function of the `img_tools` module. [#133]
 - Management of no data in cbca method. [#125]
 - Detection between false matches and occlusions. [#132]